Checkstyle for OneTopLevelClass and testsources
