+++ janusgraph-core/src/main/java/org/janusgraph/diskstorage/StoreMetaData.java
@@ -68,20 +68,20 @@ public interface StoreMetaData<T> {
             return md.isEmpty();
         }
     }
-}
 
-/**
- * Time-to-live for all data written to the store.  Values associated
- * with this enum will be expressed in seconds.  The TTL is only required
- * to be honored when the associated store is opened for the first time.
- * Subsequent re-openings of an existing store need not check for or
- * modify the existing TTL (though implementations are free to do so).
- */
-enum TTLImpl implements StoreMetaData<Integer> {
-    INSTANCE;
+    /**
+     * Time-to-live for all data written to the store.  Values associated
+     * with this enum will be expressed in seconds.  The TTL is only required
+     * to be honored when the associated store is opened for the first time.
+     * Subsequent re-openings of an existing store need not check for or
+     * modify the existing TTL (though implementations are free to do so).
+     */
+    enum TTLImpl implements StoreMetaData<Integer> {
+        INSTANCE;
 
-    @Override
-    public Class<? extends Integer> getDataType() {
-        return Integer.class;
+        @Override
+        public Class<? extends Integer> getDataType() {
+            return Integer.class;
+        }
     }
 }
